fix: convert expected_automation_percentage value to float
Error: StreamlitMixedNumericTypesError when value was int but min/max/step were float
Fix: Changed value=int(...) to value=float(...) to match other parameters

+++ pages/2_Novo_processo.py
@@ -168,7 +168,7 @@ with st.form("roi_form"):
             "% do Processo que SERÁ AUTOMATIZADO",
             min_value=0.0,
             max_value=100.0,
-            value=int(default_automation * 100),
+            value=float(default_automation * 100),
             step=5.0,
             help="De 100% do processo, qual porcentagem será possível automatizar?"
         )
